backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
langchain==0.1.0
langchain-openai==0.0.2
langchain-community==0.0.10
openai==1.6.1
google-generativeai==0.3.2
pyttsx3==2.90
scikit-learn==1.3.2
pandas==2.1.3
numpy==1.25.2
pydantic==2.5.0
python-multipart==0.0.6
python-dotenv==1.0.0
redis==5.0.1
sqlalchemy==2.0.23
asyncio-mqtt==0.16.1